updated database table

+++ sql/sql.sql
@@ -6,8 +6,8 @@ CREATE TABLE record (
   id int AUTO_INCREMENT PRIMARY KEY,
   user_id int NOT NULL,
   origin_currency varchar(50) NOT NULL,
-  origin_value int NOT NULL,
+  origin_value DOUBLE NOT NULL,
   destination_currency varchar(255) NOT NULL,
-  conversion_rate_used  int NOT NULL,
+  conversion_rate_used DOUBLE NOT NULL,
   created_at timestamp DEFAULT CURRENT_TIMESTAMP()
 ) ENGINE=InnoDB;
